+++ test/lib/perfect_shape/test_affine_transform.rb
@@ -374,7 +374,7 @@ describe PerfectShape do
       _(affine_transform.yt).must_equal -360
     end
     
-    it 'translates affine transform by (x,y) coordinate pair Array point' do
+    it 'translates affine transform by (x,y) pair Array point' do
       xxp = 2 # point x coordinate x product (m11)
       xyp = 0 # point x coordinate y product (m12)
       yxp = 0 # point y coordinate x product (m21)
@@ -394,7 +394,7 @@ describe PerfectShape do
       _(affine_transform.yt).must_equal 48
     end
     
-    it 'translates affine transform by (x,y) coordinate args' do
+    it 'translates affine transform by (x,y) args' do
       xxp = 2 # point x coordinate x product (m11)
       xyp = 0 # point x coordinate y product (m12)
       yxp = 0 # point y coordinate x product (m21)
@@ -410,6 +410,46 @@ describe PerfectShape do
       _(affine_transform.xyp).must_equal 0
       _(affine_transform.yxp).must_equal 0
       _(affine_transform.yyp).must_equal 4
+      _(affine_transform.xt).must_equal 34
+      _(affine_transform.yt).must_equal 48
+    end
+    
+    it 'scales affine transform by (x,y) pair Array point' do
+      xxp = 2 # point x coordinate x product (m11)
+      xyp = 0 # point x coordinate y product (m12)
+      yxp = 0 # point y coordinate x product (m21)
+      yyp = 4 # point y coordinate y product (m22)
+      xt = 34 # point x coordinate translation (m13)
+      yt = 48 # point y coordinate translation (m23)
+      affine_transform = PerfectShape::AffineTransform.new(xxp: xxp, xyp: xyp, yxp: yxp, yyp: yyp, xt: xt, yt: yt)
+    
+      return_value = affine_transform.scale!([3, 2])
+      
+      _(return_value).must_equal affine_transform
+      _(affine_transform.xxp).must_equal 6
+      _(affine_transform.xyp).must_equal 0
+      _(affine_transform.yxp).must_equal 0
+      _(affine_transform.yyp).must_equal 8
+      _(affine_transform.xt).must_equal 34
+      _(affine_transform.yt).must_equal 48
+    end
+    
+    it 'scales affine transform by (x,y) args' do
+      xxp = 2 # point x coordinate x product (m11)
+      xyp = 0 # point x coordinate y product (m12)
+      yxp = 0 # point y coordinate x product (m21)
+      yyp = 4 # point y coordinate y product (m22)
+      xt = 34 # point x coordinate translation (m13)
+      yt = 48 # point y coordinate translation (m23)
+      affine_transform = PerfectShape::AffineTransform.new(xxp: xxp, xyp: xyp, yxp: yxp, yyp: yyp, xt: xt, yt: yt)
+    
+      return_value = affine_transform.scale!(3, 2)
+      
+      _(return_value).must_equal affine_transform
+      _(affine_transform.xxp).must_equal 6
+      _(affine_transform.xyp).must_equal 0
+      _(affine_transform.yxp).must_equal 0
+      _(affine_transform.yyp).must_equal 8
       _(affine_transform.xt).must_equal 34
       _(affine_transform.yt).must_equal 48
     end